requirements.txt
streamlit==1.32.2
langchain==0.2.14
langgraph==0.2.27
langchain-google-genai==1.0.6
python-dotenv==1.0.1
requests==2.32.3
pydantic==2.8.2